bug fixed, wrong font loading
The AppLoading screen was introduced to ensure that fonts are loaded correctly before the main app is rendered. Previously, when the fonts were not fully loaded, sentences were being rendered incorrectly. Now, the AppLoading screen displays while the fonts are being fetched and only proceeds to show the app content once the fonts are ready